Fixed condition for PostsList to re-render
Previously, I triggered a re-render via componentDidUpdate because I wanted changes to be reflected when edits to posts were being made. However, I realised that setting the state of the component triggered re-renders. This resulted in infinite API calls.

This commit fixes the bug by only rendering the posts when there are no posts in the component yet.