Change the implementation of blog downloads
Previously the js-download-file dependency was used, which required me to specify the encoding of the files I wanted to download. A major bug was that files downloaded cannot be edited.

This commit makes use of the html-docx-js-typescript dependency to implement blog downloads, which is easier to manage as it handles HTML to docx file conversion.